Changed to one JS and done menu
Changed the game so it has just game.js instead on multiple JS states. This is so the game can use prototypes, functions for individual sprites. For coplex enemies this means that the game can just load them in with a JSON very easily. The menu has been completed, and it uses JSON loading, atlas spritesheets, and protyping. The menu is just a warmup, using the complex elements of the coding early so I can go back to them when I need for reference. Adding back the main game soon.