Finished bathroom and wrench hitbox
Finished the collision and sprites for the bathroom. For this I also made a door and panel creation function for future use. Added p2 physics for the wrench and panel, p2 will probably be used for everything except environment collision in the future, so enemies will use a mix of p2 and arcade. Also I made a very floaty jump for the player, which may have to be changed if platforming is a little to annoying with it. Working on finishing the sprites for the rest of level 1 right now.