starting second level
Finished level 1 and am now moving on to level 2. I've done most of the second level background, just need to copy and paste most of the details. Colllision has started for the first part of the second level also. The transition between the first and second level has been complete.  The jump has been made less floaty and more reliable. The next part is creating the first enemy.