Finished enemies and health system (mostly)
After a lot of work, the enemies are complete. They have a very complex attack and retreat system that should make them very interesting to battle, especially when they have you trapped or destracted by other enemies. The player health system is finished for the most part. All that needs to be added is the helmet cracking animation in the top left, and of course a respawn system. The game is no-where even remotely close to being done, but what I have got finished is still pretty cool.